fix(auth): 解决模块之间的依赖问题
auth GenCanonicalizedResource 自己计算签名时，不依赖 type 模块的特殊类型 object bucket
query 等

所有 feature 都加载 crate::types 模块，只有 core feature 才加载 crate::types::object
和 crate::types::core 模块
